Dropdown list added to report generator
Updated the input field for floor from a text input to a dropdownlist
input. Sorting the values is very weird right now as there are issues
with using Strings for the floor column.

For example:
In a list of the following String values: 1st, 2nd, 10th

The list would get sorted to the following:
1st, 10th, 2nd