Manage Rooms Changes
We have to figure out how to add a feature of inserting images for
rooms. I have modified the "Manage Rooms" page a little bit and some of
the related DAO, Controllers, etc, so we have a pretty button for
selecting features. Which gives us space to add the "Upload Image"
button beside.